+++ opencascade-hs/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to the
 - Add `OpenCascade.GeomConvert.ApproxCurve`
 - Add `OpenCascade.Geom.BSplineCurve`
 - Add `OpenCascade.Geom.BezierCurve` methods `pole`, `nbPoles` and `isRational`
+- Add `OpenCascade.GeomAdaptor.Curve` 
 
 ## 0.4.0.0